// src/app/services/model/employeeRequestModel.ts
/**
 * Metal Casting Project API
 * This API serves the frontend with data from databases and other external services.
 *
 * The version of the OpenAPI document: v1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface EmployeeRequestModel { 
    name?: string | null;
    surname?: string | null;
    position?: string | null;
    email?: string | null;
    cell?: string | null;
    address?: string | null;
}