// src/app/services/model/paymentResponseModel.ts
/**
 * Metal Casting Project API
 * This API serves the frontend with data from databases and other external services.
 *
 * The version of the OpenAPI document: v1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PaymentResponseModel { 
    paymentId?: number;
    invoiceReference?: string | null;
    dateOfPayment?: string;
    paymentType?: string | null;
    companyRegistration?: string | null;
    proofOfPaymentURL?: string | null;
    amount?: number;
    status?: string | null;
    approvedBy?: string | null;
}